More tidy up
Change magic numbers passed to tabwriter.NewWriter() into constants.

Make a table structure containing a tableRow slice.
Add an `append()` method for the new table structure to append a
tableRow to the slice.
Add a `write()` method for the new table structure to write out the
table.

Update the code to use the new table structure and methods.