More clean up
Only initialise goodStatuses once instead of doing it every time we
enter the getNodeStatus() function.

Add more constants to eliminate magic numbers.

Remove named return in getTitleRow() as golangci-lint recommended
against it.

Go 1.22+ can use range to iterate over an integer range.
Update the for loop in tabValues() to do that.

In the main function rename the r variable to row since the linter
suggested it was too short for the scope of its usage.
It suggested this for other vars too, but they were in such short code
blocks that I chose to ignore it for those as I think the code is clear.

In getNodeStatus() I only need to check len(messages) once at the end,
so take it out of the loop.

Create a lastSplitItem() function to split a string and return the last
item in the resulting slice of strings.
Use it to clean up some of the code.

Add Makefiles