Add kubectl-p plugin
Add new `kubectl-p` plugin to list pods showing details about if they
are on a spot instance and what availability zone they are in.

Add new `GetNode()`, `KubeConfig()`, `ListPods()`, and `Namespace()`
methods to the `k8s` package.

Clean up the Makefiles for the modules which hand unnecessary `ARCH` and
`BINARY_NAME` values set.