Improvements to ssm command
Implement `put` and `delete` commands for the `ssm` command.

Define a `aws.SSMClient` function and have the SSM* functions use the
instead of having to check login details each time.

Define a `aws.SSMDescribeParameter` function for retrieving the key ID
and last modified user for SSM parameters.

Have each ssm sub-command implement a struct for handling their command
line flags.

Move the code for returning an AWS profile name for an environment into
its own `getAWSProfile` function in the command root of the ssm command.

Move the code of normalising the SSM paramater path/name when not
passed as an absolute path into its own `getSSMPath` function in the
command root of the ssm command.

Improve the usage of the cobra module.