Clean up the ssm program a bit
Define the context in the top-level main.go and use cobra's
ExecuteContext() function to make it available to sub commands via a
call to cmd.Context().

Use RunE in the commands so that I can pass errors back up to the
top-level and handle them there in the once place.

Set SilenceErrors in the commands so that errors produced are not
displayed twice. Just let the top-level show them.

Set cmd.SilenceUsage via the PersistentPreRun field in the root command.
This saves have to set it in every command's RunE section.
Doing this still allows the usage to be shown for the commands if it is
wrong, but stops usage being displayed after that if the command itself
returns an error, since usage isn't applicable at that point.