Better help and usage for the ssm command
Make use of a term package so that I can get the width of the terminal.
Write a new `util.TerminalSize()` function to use it.

Write a new `util.WrapTextToWidth()` function that can format a string
to fit nicely within a certain width, making sure that words will be
wrapped to the next line if they don't fit.

Update the `init` function in the root command to get the terminal
width.
Add a new `wrapTextToWidth` template function to cobra that calls the
new `util.WrapTextToWidth()` function.
Retrieve the help template from cobra and replace the use of the
`trimTrailingWhitespaces` template function with the new
`wrapTextToWidth` one, and set that as the new help template.
Now help messages will adapt to the width of the terminal.

Retrieve the usage template from cobra and replace the usage of the
`FlagUsages` function with `FlagUsagesWrapped` passing in the terminal
width. This makes the flag descriptions wrap nicely so that the wrapped
line is indented to line up with the start of the flag description.

Allow environment names for test and prod to have more characters after
them. This allows you to have multiple environments. e.g. test1, test2,
and so on.

Make use of a heredoc module to allow me to nicely indent strings that
are defined over multiple lines. I don't like how from the second line
onwards the strings are left justified ruining the flow of your program.

Address various lint issues; mostly around comments, but a few minor
code cleanups.