Improvements
- `kubectl-p`

Replace go routines with the use of the `errgroup` module. This allows
me to run multiple routines with a wait group still and handle the
errors. I no longer need to muck around with channels myself, nor having
to add the count of go routines to a wait group beforehand.
It's cleaner and leads to simpler code.

Removed the error handling in the defer statements since trying to
return values from deferred functions can lead to strange results.
If they error, just display it at the time.

- `ssm`

Use `strings.Builder` for a more efficient way of building up strings.

Move the line wrapping logic out of the `WrapTextToWidth` function and
into its own `WrapLine` function.

Minor fix around a width calculation when tabs are involved. Created a
`lineVisualWidth` function to make things a little easier.