Profile and Region support
Have the `aws.Login` function accept a struct where an AWS Profile and
Region can be passed in, and have it act accordingly.

Update all calls to `aws.Login` to use the new struct.

Update the `ssm` command to accept new `--profile` and `--region` flags
to allow the user to override the defaults.

Update the `getAWSProfile` function in `ssm` to use the profile set by
`--profile` before falling back to its original behaviour.
Also enhance it so that an unhandled `environment` is treated as the
name of an AWS profile to use.