More improvements and fixes
Handle decryption errors in `ssm get` and `ssm list` where you don't
have access to the KMS key that encrypted it.
The `aws.SSMListSafeDecrypt` function was added to handle this rather
than just updating `aws.SSMList` since it is slower to use.

By default `ssm list` will use the faster decryption as in most
scenarios it'll be fine. But if it throws an error, the user has the
option to run it using the new `--safe-decrypt` parameter.

Add a new `Error` field to the `SSMParameter` struct to display the
error when a value cannot be decrypted.

Sometimes `SSMParameter.DataType` can be `nil` so handle that.

The `ssm.getAWSProfile` function never returned errors, so remove that
as a return parameter and update all calls to simplify things.

Set "SecureString" as a constant in aws/ssm.go

Fix or improve some comments.