Add golock program
Create a `golock` program this is based on the following:
  https://github.com/kvz/cronlock

This is because the above hasn't been maintained for many years, and I
need something that does the equivalent but to be able to talk to Redis
servers that enforce TLS encryption.

Add `util.GenEnv()`, `util.GetEnvBool()`, and `util.GetEnvInt()`
functions to get values from the environment variables and return them
in a certain type. If the environment variables are not set, then the
default value to be returned is passed to these functions.

Add a `util.Run()` function to execute a command with an optional
timeout so that it will be killed if it exceeds that runtime.