More improvements
Document the `--grep` option to the `kubectl-p` plugin in the README.

When using the `--brief` option of `ssm list`, don't add an empty line
between each entry.

Name the `ENVIRONMENT` and `PARAMETER` argument names consistently
between all the subcommands for the `ssm` utility.

Add error checking for the `ENVIRONMENT` parameter for all of the `ssm`
sub-commands.