Improvements and updates (#19)
Add a `--no-value` option to the `ssm list` command so that you can list
parameters without showing their values.
As part of that, update the `SSMParameter.Print()` function to take a
bool parameter to skip showing the value if it is set to true.

Update the `kubectl-p` plugin so that the `--grep` option now acts as a
regular expression rather than a simple string match.

Add a `make modules` target to keep the Go modules up to date and tidy.

Update the top-level Makefile to add the missing entries for the
`golock` application.

Check for errors on the deferred `rdb.Close()` call in `golock` and the
`tw.Flush()` call in `texttable.Write()`.

Update the `make lintall` target to work with the latest version of
`golangci-lint`.

Provide a configuration file for `golangci-lint` and have it ignore the
need to check for errors from the `fmt.Fprintln` function in my code.

Upgrade all the modules and set the Go version to `1.24.2`.