Use parallel routines for AWS SSM operations (#24)
Make use the `golang.org/x/sync/errgroup` module to use go routines to
run some steps in parallel in the SSM functions in the `aws` module.

Modify the `aws.SSMList()` and `aws.SSMListSafeDecrypt()` functions so
that a call to describe each parameter is not happening within the loop
that fetches them. Instead, if the `--full` parameter was used, then
after the parameters are fetched, go routines describe each parameter
in parallel to fill in the extra parameter attributes.
This removes a major bottleneck that my previous version had.
A semaphore is used to allow up to 20 go routines to run in parallel to
avoid overloading the AWS SSM API endpoint.

Have the `aws.SSMGet()` function use go routines to perform the
`GetParameter()` and `DescribeParameter()` calls in parallel.

Add a boolean `describe` parameter to the `aws.SSMGet()` function to
define if the parameter should be described to get extra attributes.
Update all calls to the function accordingly. The `ssm get` command
passes `getOpts.full` for this new parameter so that it won't waste
resources fetching attributes that will not be shown to the user.

Capture errors from the calls to `SSMDescribeParameter()` properly.