Add --json parameter to ssm get and list commands
Create a `util.MarshalWithoutFields()` function to marshal a struct into
JSON while being able to omit some struct fields.

Update the `aws.SSMParameter` `Print()` method to accept a `json` bool
parameter to control when you want it to display JSON output.
Break the function up to use helper functions to make the code more
readable.

Add a `--json` parameter to the `ssm` `get` and `list` commands.
Pass this to the `aws.SSMParameter.Print()` method when that is used, or
use it to make sure other output is in JSON format.

Break up the `displayListParameters()` functions in the `ssm` `list`
command to use helper functions to make it more readable.

Update some Go modules.