More improvements
Have the client name used by SSO passed in by the client instead of
having it hardcoded.

Check the error during SSO token refresh and change the message if the
token is expired, vs any other error.

Add `getSSMClient()` function to the `ssm` command to reduce some code
repetition across the various subcommands.

Have the `ssm` command pass the `ssm` client name to the AWS login.

Removed a pointless initialisation of a string variable `keyID` to and
empty string, since that happens automatically when a string variable is
defined.