Add ability to display the version of the tools
Define a `version` variable in the `util` package and a
`util.DisplayVersion()` function to display it.
The `version` variable is set via `-ldflags` at build time.

Add a `VERSION` variable and `LD_FLAGS` variable that uses it, to the
`Makefile` of the `ssm` tool, and kubectl plugins and have the targets
use them when building the binaries.
The version is in a CalVer `YYYY.0M.0D` format with a git commit hash
added to the end. The date and hash is taken from the latest commit in
the branch the build is run from.

Add a `--version` flag to the kubectl plugins to call the
`util.DisplayVersion()` function and then return.
Add a `version` subcommand to the `ssm` tool to also call
`util.DisplayVersion()`.

Change the Makefiles to use POSIX `=` when assigning variables instead
of GNU Make's `:=` operator. No reason other than consistency, and I
need the recursive nature of the `=` operator when evaluating the new
`LD_FLAG` variable so that the `VERSION` in it is expanded too.