Improve SSO login
Generate and use the `state` query string in the SSO Login URL.
It is a Cross-Site Request Forgery protection mechanism that should be
there.
It is used to validate that the response from the AWS OAuth provider
corresponds to the same request our application made.

Rename `aws.startLocalCallbackServer()` to `aws.localCallbackServer()`.

Modify the `aws.localCallbackServer()` function to accept the expected
state as a parameter and to verify it.

Add a context parameter to the `aws.localCallbackServer()` function and
use it when creating the listener.

Change the callback URI to include a `/oauth/callback` path for no
reason other than to match what the AWS CLI uses.

Upgrade some more Go modules.