Set ssm default completion option
Upgrade Go modules.

The new spf13/cobra module allows you to set the default completion
option, so take advantage of that to stop the file completion helper
from being the default, since none of the options for the `ssm` commands
except for `put --file` want a filename to be passed to them.

Explicitly set the `put --file` option completion function to use the
default of using standard shell file completion.