Change ssm list to use one line format by default
The `--brief` option is gone and is now the default output since that's
how I find I used the tool all the time.

Add a `--verbose` option that produces what used the be the old default
multi-line output.

Update the README.md file since the usage examples were out of date.

Update the version of go and go modules.