More kubectl-p options (#39)
Define a new `k8s.PodInfo` struct and have the `k8s.PodDetails()`
function return that instead of a bunch of separate variables.

Rename the `--grep` option to `--name` for `kubectl-p`.
Add new `--node`, `--not-name`, `--not-status`, and `--status` options
to `kubectl-p` for better ability to filter the pods that are matched.

When processing `--name` (previously `--grep`), compile the regex once
instead of within the function passed to `slices.DeleteFunc()` that
would end up compiling the regex for every pod that is evaluated.

Minor simplification in the `texttable.Write()` function.

Update the README.md files for kubectl-n and kubectl-p to match their
new options.

Update modules.